Blink on update
When we query the weather API, give the user some feedback through the bulb that
the service is still up and working. At each `INTERVAL_SECONDS` value, we'll
blink the bulb.

We may also in the future have an update blink interval when we know the score
will change, based on the forecast.